Frontend: added notification & dialog for tag actions

+++ src/frontend/src/app/templates/admin/admin-tags/admin-tags.component.ts
@@ -3,6 +3,10 @@ import {ActivatedRoute} from "@angular/router";
 import {TagService} from "../../../services/tag.service";
 import {FormControl, FormGroup, Validators} from "@angular/forms";
 import {Tag} from "../../../models/Tag";
+import {MatDialog} from "@angular/material/dialog";
+import {DialogComponent} from "../../../shared/dialog/dialog.component";
+import {MatSnackBar} from "@angular/material/snack-bar";
+import {SnackbarComponent} from "../../../shared/snackbar/snackbar.component";
 
 @Component({
   selector: 'app-admin-tags',
@@ -15,7 +19,8 @@ export class AdminTagsComponent implements OnInit {
 
   public tagForm: FormGroup;
 
-  constructor(private activatedRoute: ActivatedRoute, private tagService: TagService) { }
+  constructor(private activatedRoute: ActivatedRoute, private tagService: TagService,
+              private _dialog: MatDialog, private _notification: MatSnackBar) { }
 
   ngOnInit(): void {
     this.tags = this.activatedRoute.snapshot.data.tags;
@@ -38,8 +43,30 @@ export class AdminTagsComponent implements OnInit {
   }
 
   delete(tag: Tag) {
-    this.tagService.deleteTag(tag.id).subscribe(() => {
-      this.tags = this.tags.filter(t => t != tag);
-    }, error => console.log(error));
+    this._dialog.open(DialogComponent, {
+      data: {
+        title: 'Подтвердите действие',
+        message: `Вы уверены что хотите удалить тэг "${tag.name}"? Возможно, есть идеи которые его используют.`
+      }
+    }).afterClosed().subscribe((result: boolean) => {
+      if(result) {
+        this.tagService.deleteTag(tag.id).subscribe(() => {
+          this.tags = this.tags.filter(t => t != tag);
+          this._notification.openFromComponent(SnackbarComponent, {
+            duration: 2000,
+            horizontalPosition: 'left',
+            data: 'Тэг успешно удален!'
+          });
+        }, error => {
+          console.log(error);
+          this.tags = this.tags.filter(t => t != tag);
+          this._notification.openFromComponent(SnackbarComponent, {
+            duration: 3000,
+            horizontalPosition: 'left',
+            data: 'Произошел сбой, попробуйте позже.'
+          });
+        });
+      }
+    })
   }
 }
